Update RunJobForm.php
Added VirtualFull option for manual defined Jobs. In my case, sometimes you have a Full Job and some Incremental  ones and you just one to merge them into a new Virtual Full. From the UI it is really simple to do it if we add this option to the menu.