dird: display who triggered the job in job reports
This commit introduces an enum class named JobTrigger used in
JobControlRecords. The purpose is storing and tracking down by whom a
job has been triggered and to display its value in job report summaries.

Currently the following values are defined, kScheduler, kUser and
kClient. The default is kUnknown.

The commit also adjusts all the different job summary generators
to display the JobTrigger in a human readable format.