oVirt plugin: fix restore problem to local disk
Fix local restore problem introduced with commit 112020886099.

During local restore, the restore is cancelled with a KeyError:

    File "BareosFdWrapper.py", line 66, in create_file
      return bareos_fd_plugin_object.create_file(context, restorepkt)
    File "BareosFdPluginOvirt.py", line 268, in create_file
      context, restorepkt.ofname
    File "BareosFdPluginOvirt.py", line 1521, in get_ovf_disk_alias_by_basename
      return self.ovf_disks_by_alias_and_fileref[relname]["disk-alias"]
  KeyError: ('testvm1_thin_cow-3b0d105b-0803-4964-be21-ede95153761c/524941fd-ef79-494f-9831-1d49cff9ed76',)

The cause is that the disk alias filtering uses information that is not
fetched from the oVirt Server during local install, as it is not
contacted at all.

During local install, no disks are filtered and every disk backed up will be
restored.

Fixes #1246: oVirt plugin fails to restore to local disk with KeyError