lib: avoid deadlock in signal handler
Previously the signal handler allocated memory by calling
GetPoolMemory() indirectly via the BErrNo ctor.
In case the signal handler was called while the mutex for pool memory
was locked, the result was a deadlock.
This patch removes the usage of BErrNo in favor of standard strerror()
which avoids GetPoolMemory() and in turn fixed the deadlock.