cmake: fixed "make install" and added "make uninstall"
The "make install" target created by cmake was not completely working.
Also, "make uninstall" was not available.

This patch fixes the problems with the "make install" target and added
the "make uninstall" target by reading the installed files from the
cmake manifest file and deleting them.

The code for the uninstall target was taken from:
https://gist.github.com/royvandam/3033428