plugins: enable restore of plugin data with newer plugin
When trying to find the right plugin for an event in "IsEventForThisPlugin()",
not only the full plugin name is matched like before.
Now also we check if the name without the last character Matches.

This is helpful to be able to restore data written with the "python"
plugin with the "python3" plugin.