filed: ldap - detect missing data during restore
When the LDAP plugin fails to back up an object, it will be recorded in
the catlog, but the data will be missing.
Previously these objects were silently ignored during restore, leading
to an incomplete set of restored objects.
This patch now adds a rudimentary detection of this case by checking if
plugin_io() is called with IO_READ, IO_WRITE or IO_SEEK between the
calls to IO_OPEN and IO_CLOSE. When there is no call we now emit a
warning that there was no data to restore the object.

(backport of commit e9b60b1b7c320abcda2288ffa2c394880a5858cd)