filed: ldap - detect slashes in DN
Bareos cannot handle objects with a slash in their name gracefully,
because it is the path delimiter.
This patch adds detection of DNs the contain a slash to the plugin. It
will now correctly raise an error and will no longer try to back up that
data.

(backport of commit 0f53026cba7a316b5baa38c9aca51437b0782fa9)