core: inherit class db_list_ctx from BStringList
This change improves the handling of output from the "delete" command.
The changes are required to access the individual jobids in the list,
instead of having only access to the a string containing the comma seperated list of jobids.

The count() function has been replaced with the standard size() function.
The reset() function has been replaced with the standard clear() function.

cleanup

cleanup