core: prevents a director crash when calling "configure add schedule"
The code to parse a schedule run directive calls the scan_err function
in case of a parse error.
However, it did assume, that scan_err terminates the director.
While this is true when parsing at the start of the director,
it isn't true, when called by the "configure add" command.

This patch improves the error handling, by exiting the parse function
when an error is detected.