core: configparser: cleanup reading CFG_TYPE_ALIST_RES
This commit cleans up the configparser when handling CFG_TYPE_ALIST_STR directives.
This code did contain legacy artifacts, misusing the default value to address an array of alists.
As none of the directives using this type have a default value, it always addresses the first alist
and therefore can be treated
in the same manner as CFG_TYPE_ALIST_STR or other configuration list directives.