docs: clarify FileSet->Include->File directive (use of backslashes)
Before, the documentation shows \<includefile-client and mentioned,
that for quoted strings, two backslashes must be used.
However, the configparser does not allow backslashes in File directives in unquoted strings,
therefore two backslashes must always be used.
Moreover, when using a single backslash in a quoted string,
the configparser checks the next character.
In case of a <, it just ignores the backslash,
so that File="\<includefile" is identical to File="<includefile".

With this change, the documentation only shows the working format: "\\<includefile".

The corresponding change is also done for File = "\\|command".