core: don't assume, that scan_err*() terminates the program
If the Bareos parser finds an error in the configuration,
it calls scan_err*().
In most cases, scan_err*() terminates the Bareos daemon.
To document this, the comment "/* NOT REACHED */" has been added
after some of the scan_err*() calls.
However, the parser can be configured to return from the scan_err*() function.
In fact, this is done when called by the console "configure" command.
For this reason, the "/* NOT REACHED */" comments have been removed,
as they are not always true.
Instead "return" is called after scan_err*().