python-bareos: fixes support for sslpsk with Python >= 3
The current sslpsk version requires identity and password as bytes.
Strings are no longer supported.  python-bareos is now adapted to this.

Unfortunately, depending on the Python and OpenSSL versions
sslpsk (1.0.0) does not work with all TLS protocol versions.
Normally, the best matching protocol version is selected automatically,
but this fails in many environments using sslpsk.
As a workaround it is now possible to specify the TLS protocol version (v1, v1.1, v1.2)
python-bareos should use.
Note: newer protocols (TLS v1.3) cannot be specified in this way.