regress: make sure python-template is valid python
previously the template used @dirport@ as an integer, which is
considered a parser error. We now wrap the value with int("@dirport@")
which will make sure the template is valid python code and will throw a
runtime-error during conversion instead of a parser error it @dirport@
is replaced with anything that is not an integer.