vmware: check VDDK bug triggers in wrapper script
The VDDK after 6.5 has at least two issues that can cause a crash,
this patch tests for the circumstances that trigger the bug and
will lead to a crash, telling the user how to work around the issue.

The patch also removes the setting of LD_LIBRARY_PATH as it is not
required anymore.