Remove packages bareos-regress-config and bareos-devel
bareos-regress-config has only been a helper package for the bareos-regress package,
which is currently not build.

bareos-devel has been incomplete for a while. Therefore it seams, it is not really in use and can be removed.

freebsd pkg-plist.traymonitor file is removed, as it is not build anyway.

removed docs/pkglists for EOL distributions:
Fedora_30, RHEL_6, SLE_12SP4, openSUSE_Leap_15.0, openSUSE_Leap_15.1