plugins: add Debian packaging for PostgreSQL plugin
Add Debian packaging for bareos-filedaemon-postgresql-python-plugin.
python*-dateutil is required for bareos-filedaemon-postgresql-python-plugin.
Adding the requirement to the packages bareos-filedaemon-python2-plugin
and bareos-filedaemon-python3-plugin,
as these handle the different type versions
and it is also required by other Python modules (e.g. libcloud).

Pluging packages prefer Python 3 instead of Python 2.
While is is posible to install both plugins,
it is not possible to load both at the same time into the filedaemon.