bstrncpy: workaround when used with overlapping strings
Workaround for a problem, that has occured
while parsing a configuration file containing a long time type directive.
However, it has only be observed on
  * Debian 9, 32bit,
  * CXXFLAGS contains -fstack-protector-strong (default for Bareos packages)

In this situation, bstrncpy(str, &str[NUM], len) does not work as expected.
Instead, only the initial bytes from str[NUM] are copied, but not the full string.

This workaround tries to detect overlapping strings
and uses a copy instead.

Unittests have been added to test the behaviour.