systemtest config-dump: ignore whitespaces in config files
Workaround for a problem introduced by 84a44ca6cc4d3bd4fcbe428c6433722cae251d35.
This test has been whitespace sensitive, but that commit removed all trailing whitespaces.
As a workaround for this problem, we use
diff --ignore-blank-lines --ignore-space-changes
instead of a diff without parameter.
The parameters of diff and sed have been tested to work on Linux and on FreeBSD.