systemtests: set -o pipefail
While the execution of a command with exit code other than 0 is normally an error.
The test scripts stop than with status error.
However, when commands are executed with pipes,
only the exit code of the last command is evaluated.
With the option "pipefail", the first failing command sets the exit code.
This is useful, because otherwise errors can occur unnoticed.