core: make "alist" and "dlist" type safe templates
The two data structures "alist" and "dlist" that are used widely in the
code are not type safe as they only handle void* pointers that are
cast-ed from and to the required datatype manually. This is both
error-prone and leads to code that is hard to understand.

This commit changes the two types to templates that are type safe.

As the usage of dlist in the chunked-device backend was totally misused
so that the change to the template dlist was not possible, the old dlist
still exists there as "old_dlist" which needs to be removed in a later
step.