dev.cc: do not call virtual functions in destructor
The destructor of the the Device base class can not call
cannot call close(), as close() itself calls virtual functions
but the destructors of the derived classes already have been called before
so the overridden functions are not available anymore.
Calling virtual functions of a class need to be called in the destructor
of that class and not of the base class.

Add a unittest for testing this (load and unload tape device).

Co-authored-by: Frank Ueberschar <frank.ueberschar@bareos.com>
(cherry picked from commit 6a10326c9e0d809f19a38a3283c97c5b88cb1db3)