stored: reorder acquire device for migrate/copy
previously migrate/copy would acquire the device for reading before the
device for writing. This could lead to deadlock situation when the
volume that should be read was mounted in the device reserved for
writing.
By acquiring the device for writing first any volume in that device will
be unloaded unless it is the volume we are going to write to. This
should fix the deadlock described above.