stored: remove old_dlist/old_dlink
Because of weird usage of dlist and dlink in ordered_cbuf there was a
copy of the original dlist/dlink as old_dlist/old_dlink. As this is
basically the same as using the templated version with typename void,
this patch replaces all uses of old_dlist/old_dlink with dlist<void> and
dlink<void> rendering old_dlist/old_dlink unused.