systemd: switch service from forking to simple
The service type forking caused problems on newer systemd versions
(Fedora 34).
Using type=simple and starting the daemons without forking
has the advantage, that daemon messages to stdout and stderr are shown
with systemctl status or journalctl.  This includes the daemon debug
messages, when set into debug mode.

Removed the PIDFile directive, as it is only required for Type=forking
and will be incorrect, as soon as listening port is changed in the
Bareos configuration.

Removed the no longer supported (and never really used) directive
StandardOutput=syslog

Director: Exit code 1 is no longer treated as success.  This has been
introduced as a workaround, but is no longer required.