vmware: Fix bareos_vadp_dumper VMDK File creation
Before, the value of length attribute of the changed block tracking query
result was used as disk size when bareos_vadp_dumper called the VDDK function
VixDiskLib_Create to create VMDK files on restore with localvmdk=yes. But
this value is not the real disk size in all cases, so the VMDK was too
small.

With this fix, the VMDK files are created with the correct size.

(cherry picked from commit 1c190f51720ff0b1208df6b07c17b1f28bfcddab)