tests: fix data-race in statefile test
previusly the read and write test used to same statefile path which lead
to a race on the file if both tests were run at the same time (as is the
default when using `ctest -j X`.
This patch makes the read-test use a differently named file so there is
no race anymore.