dird: changed the position of volume list memory allocation
When in the while loop, the `vl` variable gets allocated some memory
AFTER the fields check, and then this memory gets freed later in the
loop. But when there is an iteration that stops at the check and gets
redirected to `parse_error`, it has to free memory that was already
freed in a previous iteration of the loop, which creates a double free
situation that crashes the director.

(cherry picked from commit 65c1185bcf5236546f50789c0859a4dd89b9144c)