docs: remove styleguide for referencing a release
This will not work reliably with the newer CHANGELOG.md as all the
referenced versions will be gone as soon as a new major release is done.
In the current documentation there are no references to specific
versions that are done in this way, so we can probably safely remove
this.

(cherry picked from commit 4d679b40c7102605ac8b0bc5e2692a215fc14e8f)