systemtests: configurable database credentials
This changes makes the PostgreSQL database configuration
for systemtests simpler.
Before it has be necessary to configure two database roles,
one for the current system user and one for the database user "regress".
The database admin scripts did run as the system user,
while the Bareos use the "regress" database user.
This did also require to adapt a PostgreSQL configuration file.

With this change, only database one role and no change
to the PostgreSQL default configuration is required.

For using the system user, run cmake with following parameter:

cmake -Dsystemtest_db_user=$USER ...

When systemtest_db_user is not given as parameter, the old default "regress" is used.