doc: fix handling configuration directives
The Bareos configuration directives are expected to be written in CamelCase.
It turned out, that this is not always the case,
resulting in mismatch of documentation files and broken internal links.

This changes renames some configuration directives into proper CamelCase.
It also tries to convert directives into proper CamelCase,
adds link targets also for :config:option: with 2 components
and produces warnings on dangling links,