lib: fixed empty error messages on windows
certain windows errors do not have a message to accompany them!
The `formatMessage()` function returns an empty string when called for an error
that does not have a registered message. Basically, not all error strings are defined
in the system message table.
So the change that was made was to make sure that if the error message does not
exist, at least the error code is returned to the user so they can search for the
appropriate error reason somewhere else