lib: show windows error codes along with unix errors on windows
The errors would not use the necessary windows code when `b_errno_win32` 
is not set. And so, the program automatically tries to search for a Linux 
interpretation of the error code, which would lead to a wrong/non-existant message
for that error code. Windows and Linux error codes are different and only
have a few common error codes like `0` for `success` or error `2` for `no such file
a directory`.
This change makes sure that, when in doubt, just show both messages and let the user interpret the error code provided.