Wael_Youssfi Assignment
*First task (the output): I sent the output as plain text to the email: jobs@bareos.com
*Second task (How  to improve setdebug behavior): 
When using the full command of setdebug level=100 dir trace=1 for the second time the file is created containing the event or the trace should I say.
However, when using the setdebug interactively it doesn't ask for the trace flag (0 or 1). After testing out a couple of times, it uses the last provided flag in the previous full command whether you provided 0 or 1.
Example: if I use setdebug level=100 dir trace=0 then try to use it interactively the trace file won't be created in case I deleted it after running the first full command because the interactive setdebug would have for a trace flag 0 in this case.
How to fix in my opinion: Quite simply I believe that we need to prompt for the user to ask him whether he wants trace= 0 or 1 in the same way we ask about the level 
*Third task (Identifying the codebase function): 
1)In the code-base the function that handles this functionality is in the file: bareos/bareos/core/src/dird/ua_cmds.cc specifically line 1397 "static bool SetdebugCmd(UaContext* ua, const char* cmd)" .
2)The function is called in line 205 in the file bareos/core/src/stored/dir_cmd.cc and in line 182 in the file bareos/core/src/filed/dir_cmd.cc line.
3)The SetdebugCmd function calls mainly the following functions: 
DoAllSetDebug
DoDirectorSetdebug
DoClientSetdebug
DoStorageSetdebug
*Fourth task (A way to implement the new behavior):
One of the ways to implement the behavior I mentionned in the second question is to:
To prompt the user and ask him to input the trace flag value 0 or 1 
This will be by adding some code in the implementation of the SetdebugCmd function 
*Fifth task (View the content of the pull request)
*Sixth task (Other unexpected behaviors):
Firstly, when I was testing out the setdebug command in the bconsole, I noticed that the trace file will not be created from the first call of the setdebug command but it will be created after the second call.
Secondly, sometimes when I delete the created trace file and execute the setdebug command the won't be created even if I execute the command for many times (with trace= 0 or 1 no difference).