cmake: add targets cov-report and cov-clean
This patch adds two new custom targets to CMake if building with
coverage enabled.

cov-report:
  will generate a coverage report in CMAKE_BINARY_DIR/coverage

cov-clean:
  will wipe out the coverage report cov-report generates and will also
  remove the gcda files in the build tree resetting the collected
  coverage information.

Previously the ctest pretest executed the director binary which would
add misleading information to the coverage report (i.e. director code
was run wheter or not you were running a director test). This patch also
changes the pretest to just check for the executable instead of calling
it.