dird: fix for 'current' option wrong output and other issue
Setting ACL filters and resource filters requires knowledge of the 
exact position of the query columns. Since the query was updated
to show duration which was put just in the middle after starttime
and endtime, the subsequent columns positions were incremented
which rendered the code obsolete.

So this commit updates the positions of the required columns