lib: fix undefined behaviour in ITEM() macro
Previously ITEM(c,m) was called with `c` being a NULL-pointer with the
type of one of BareosResource's children. It then took the address of
`c->m` as the offset of m within the type of `c`. This is undefined
behaviour. It also relies on `c == NULL`, which was never checked.

This patch replaces the existing method by using offsetof() with the
non-pointer type of `c` and its member `m`, removing all NULL-pointer
involvement and allowing to pass any pointer or non-pointer value of the
desired type.