lib: remove pointer-artihmetic from hlist<>
Previously hlist<> got two pointers on construction that were used to
determine the offset of T::link at runtime. As this is now done at
compile-time, the pointer-arithmetic is not needed anymore and
construction doesn't need the pointers anymore.
This patch removes the pointers, finally removing the undefined
behaviour from hlist.