build: stop building Debian 9
the Debian Project will EOL the LTS version of Debian 9 on 2022-06-20
which is way before the planned release of 22.0.0. As we will not build
Bareos 22 for Debian 9 we can disable it now. As Debian 9 was the last
system that did not have a compiler with full C++17 support, the
immediate benefit is that we can now finally use C++17.

We will also disable Univention 4.4, which is built on Debian 9.