gtest: fix leaks related to message initialization and deletion
when initializing messages using InitMsg, message resources
are created in `message.cc` but not cleaned up afterwards.
One solution could be to provide the tests with a `TermMsg`
in order to cleanup, but it looks like all those `InitMsg`s
do not have any effect on the tests. So this commit will delete
all unnecessary occurrences of `InitMsg`