refactoring lockmanager name to avoid conflicts with CLI11
Macros `P(x)` and `V(x)` create conflicts with CLI11.
We decided to remove the macro all together and rename the 
function to be more expressive.

Anecdote: 
V stands for 'Verhoog', which can be translated as "increment" from dutch.
P stands for 'Prolaag', a madeup Dutch word of 'probeer verlaag',
which can be translated as "try to decrease".

These words were coined by Edsger W. Dijkstra.