docs: add build target docs-check-urls
Use the command

make docs-check-urls

to verify that all used links in the documentation are still valid.
Exceptions can be configured in docs/manuals/source/conf.py

As this command can be time consuming and error prone (due to internet or server problems),
it is not configured to run automatically.