dird: add extra 6-character truncation to anticipate psk identity prefix
When creating a Job name, we do a truncation to add a suffix and
make sure the string does not pass the 128 character limit.
Later, during TLS-PSK authentication process, a prefix `R_JOB^` is
added to the job name to create the identification key, and another 
truncation needs to happen in order to comply with OpenSSL1.1's 
128 character limit for identity.
This second truncation eats the last part of the Jobname containing
our custom suffix, and thus creates problems.
This issue does not show up with OpenSSL3, as the identity limit was
changed to 256 characters.

This commit then anticipates the extra 6 characters needed for the 
"R_JOB^" prefix added in the psk identity string.