systemtest: added systemtest for untested CLI options
Most of the bareos CLI options are used throughout tests and are 
therefore tested, meaning if there is any issue with one of those options, 
the tests that use them will break. This is valid for -f, -d, --dt, -c, --xs, etc.

Certain other options like -u, -g, -r, -p are not tested, and there is no 
basic way to test if at least the daemons don't crash because of them.