VMware plugin: Fixed: Restore to local file required API connection
Restoring a VM to a local file using the VMware plugin required a valid
connection to the vSphere API.
This makes a restore in a worst-case scenario where you lose your
vCenter installation unnecessarily complicated.
This commit fixes the option check logic and does not connect to the
vCenter API when localvmdk is set to yes.