win32: refactor all uses of GetProcAddress
This patch wraps GetProcAddress into a C++ template to make it easier to
use and to allow us to ignore the cast-function-type warning in a
central location.
It also simplifies a lot of the uses due to template argument deduction
and changes winapi.cc to use macros that reduce repetition.