python module: improve field initializers
When initializing PyTypeObject, there are a lot of fields to set. To
avoid problems due to mis-ordering, we change this to use designated
inizializers. While these are not legal in C++17, they're in C99,
available as a compiler extension on virtually any major compiler and
will became legal in C++20.