plugin api: transfer FO_XATTR and FO_ACL from savepacket to core
The XATTR and ACL file option bits were not transmitted to the core for
evaluation.

When the savepkt->flags are set to FO_XATTR and FO_ACL, this information
was not transferred back to the core and thus the get_acl() and
get_xattr() calls were not emitted to the plugin.