core: distinct names for JobControlRecordPrivate
When we moved the daemon-specific functionality of the JobControlRecord
into JobControlRecordPrivate, we essentially introduced three different
types that were all named JobControlRecordPrivate - one for the
Director, one for the Storage Daemon and one for the File Daemon. This
could lead to ODR violations when you tried to link parts of the daemons
together.
In fact, that implementation already behaved like a union of the three
different types (i.e. requiring per-implementation initialisation and
teardown, undefined behaviour when accessing via the wrong type, etc.).

This patch renames the daemon's individual JobControlRecordPrivate to
DirectorJcrImpl, StoredJcrImpl and FiledJcrImpl. The impl-pointer was
changed to a union of dir_impl, sd_impl and fd_impl.
With this change applied you can now build programs that use two or more
of the JCR types.