restore.cc: file count mismatch leads to job warning
A file count mismatch during restore only changed the result message,
but not the job term code.

Now the job term code is changed to warning, and now also this is issued
if the number of files restored are **different** to the number of files
expected.
Before, the warning was only issued if **less files than expected** were
restored.

Co-authored-by: Philipp Storz <philipp.storz@bareos.com>