rpmbuild: suggest bareos-*-python3-plugin
Previously the RPM depsolver selected the python2 plugins. So if you
installed bareos-filedaemon-postgresql-python-plugin that would require
bareos-filedaemon-python-plugin which in turn was resolved by adding
bareos-filedaemon-python2-plugin to the transaction.
With the suggests added with this patch, the depsolver will not prefer
the python3 plugin, so that in the above example
bareos-filedaemon-python3-plugin would be installed.

As rpm on RHEL 7 does not yet support suggests, the fix does not apply
to that platform.