devtools: support multiline records with add-changelog
add_changelog_entry() did a bad job if there were existing multiline
entries in the changelog. If the section where a new entry should be
added ended in something like this:

- normal changelog entry
  - with additional information
  - and more information

it would misdetect the location to add the new record and insert it
after the first line instead of at the end.
Now it also skips lines where "- " is preceded by whitespace.