devtools: fail pr-tool when GH_HOST or GH_REPO set
The user can override the host or repository for the gh command using
the environment variables GH_HOST or GH_REPO. As we will have to
determine the repository and the name of the remote ourselves (gh does
not provide a way to do that), setting GH_HOST or GH_REPO will probably
break pr-tool at some point, so we disallow it.

Also introduces logging options.