devtools: pr-tool warns if remote not uptodate
When your local repository is not up to date, rebasing (and potentially
other operations) will not lead to the desired result or may even fail.
pr-tool will now check if your local repository's base branch points to
the same commit as the remote repository's base branch and will issue a
warning otherwise.