devtools: remove mime from bareos-check-sources
Instead of looking at the mime-type, we now simply check if the file is
empty and try to read it as utf-8 otherwise. This turned out to be
faster than guessing the mime-type first.
We can probably improve this further by adding some filename patterns
that should always be skipped (like *.gz, *.tgz, *.png, *.min.*).