stored: fix broken volume swapping
Fix a problem introduced in Bareos 22 with commit
6c265ec52 replace volatile device pointers with actual variable

Due to an oversight in a refactoring, a swap would no longer move the
volume to another device, but switch over to use the device that had the
desired volume loaded.
As that device was not reserved, but was unreserved during acquire, the
storage daemon would eventually crash in this case.