cats: stored: only create jobmedia table entries
Previously, when updating backup checkpoints, DIR would try to
update the jobmedia table entries or create a new one in case 
an update is not possible which in consequence made the 
jobmedia table entries kind of merged.

The change lowered the number of entries in the jobmedia table
but had the unwanted consequence of slowing restores in certain
situations (restores on big volumes for example).

With this change, jobmedia entries will only get created, and 
checkpoints will add an entry into the jobmedia table instead
of updating existing entries.