bugfix: stored_conf: Fix validating wrong resource
During pass 2 dummy resources are being created that end up not being
really used.  These dummy resources often do not get filled with the
complete information; in particular device_type was never filled out
since its std::string.

Previously the dummy resource was being used to validate the resource
configuration.  This lead to spurious warnings since the validator
expected the data to be actually correct, e.g. it checked to see if
the device_resource was a tape or not.  Since this info was not filled
out it threw a warning.

Now the resource created in pass 1 is used for validation.