gtest: add warning to disable local dir when failing
If you have a local bareos director running that is using a port that
a test wants to start a director on it will fail.  Normally this will
not happen but one particular test checks the default config.

In that case it might not be obvious to check wether you have a bareos
director running in the background.  This hint was added for this
particular purpose.