find: make ListFiles return number of skipped files
To correctly set the number of files examined, we need to know the
number of files that SaveFile looks at (this is the number of files in
the various channels) and the number of files SaveFile does not look
at (i.e. the number of files that ListFiles skipped).

On success ListFiles now returns the number of files it skipped.  We
use that to correct num_files_examined.