bfile: better init value of BareosFilePacket
We should not use a valid fide descriptor on a bfd that was not
initialized yet.

Because of this IsBopen(bfd) curretly returns true (on unix systems)
if bfd was a newly created fd on which binit() was not called.
On windows systems it already correctly returns false in that
situation.  Now it always returns false.