systemtests: add function systemtest_requires()
Introduce a way to express hard dependencies between systemtests.

Using DEPENDS on a test only ensures ordering, but does not imply the
need to run the other test or that the other test must have finished
successfully.

Using the new function systemtest_requires() you can now express that a
test A will require successful execution of test B by simply calling
systemtest_requires(A B). As this is tailored to be used inside a single
directory, the required prefixes are added automatically. The function
also honors existing fixtures.
Under the hood this is implemented by adding a setup fixture to every
systemtest that is then required by the depending test.