dird: fix use-after-free after a failed reload
When reloading the configuration failed, the working_directory would not
be reset correctly which made it point to the freed string of the new
configuration.
Now SetWorkingDirectory() is called after reloading in case of success
and failure.

This also allows to run all of the reload systemtests under a sanitizer
which failed previously because of the use-after-free.

(cherry picked from commit 91f35fa5afcd5caa5c6e59c60c84878c776adabf)