webui: add a new section to the `configuration.ini`
Adds a new section named `experimental` to the `configuration.ini`
where experimental features can be enabled and disabled.

All experimental features are and should be disabled (`false`)
by default.

First item in this new section is the configuration resource graph,
which is located in the analytics module and reachable via the
navigation tab `Configuration`.

The navigation tab `Configuration` is hidden if the newly introduced
`configuration.ini` parameter `configuration_resource_graph`
is not set or set to `false`.