channel: give ability to wait for free space
Now producers can wait until some percentage of the queue is free
instead of waking up directly once exactly one spot is free.  This is
useful for cases where the producer is much faster than the consumer
since we do not want to constantly wake up a thread only to do just
one unit of work and then go to sleep again.