catreq.cc: handle requests to reduce VolFiles
We have the effect that at the end of restores, we get catalog requests
that try to reduce the VolFiles.
Instead of emitting a fatal error, we now log the event and keep the
current value.

We do the same for VolBytes and for VolBlocks.

(cherry picked from commit 270d97169b83b0995839b9fefc884b3d07104063)