webui: make restore merge options configurable
If you select a client in the webui restore module the default is that
according jobs and filesets get merged for the file tree presentation.

In larger environments you probably do not want this default behaviour.

Therefore, this commit introduces two new configuration.ini settings
to switch the default merge option behaviour in webui restore.

The newly introduced configuration.ini parameters are "merge_jobs"
and "merge_filesets" in the restore section.

Co-authored-by: Joerg Steffens <joerg.steffens@bareos.com>