compat: replace thread local variable by global variable
There is zero reason for this variable to exist at all let alone to be
thread specific.  It literally only contains two function pointers
that are always the same and no thread specific data.
For now, if no thread specific variable was set up,
we just return a global variable if it was set up.

This should enable non-main-threads to "interact" with the vss client;
i.e. look up drive -> vss drive translations (which are readonly!).