find: remove sys_fname+ignoredir_fname from ff_pkt struct
These are only used in one function each so they can just be converted
to local variables.  Reducing the size and complexity of ff_pkt is
worth the 2 extra allocations per file.  There are better ways to
reduce these kinds of allocations (if they become a problem) than to
just stuff them in the already-to-big ff_pkt struct.