VMware Plugin: improve snapshot cleanup, bootOrder
If leftover snapshots exist from previous backups, they are now cleaned
up before taking a new snapshot. Additionally, the snapshot cleanup
after backup now works more reliable by reconnecting to the vCenter
server if necessary. Code refactored to consequently use methods from
pyVim.task.

- fix a bug when folder=/ is used.

Fixes: #1533: Restoring vmware vms keeps failing, can't restore the data

- Fix restore when bootOrder was set

When bootOrder was set to a virtualDisk at backup time, ensure the
restore works by setting the bootOrder now after disks were added to
the recreated VM.

- Fix restore when virtualCdrom path is invalid

Restore fails when recreating a VM and the backing file of a virtual
CDROM is inaccessible, this can occur when it was connected to an ISO
file on a NFS datastore which is not accessible at restore time.
This change will then create a disconnected virtual CDROM to allow
the restore to finish successfully.

Co-authored-by: Philipp Storz <philipp.storz@bareos.com>