config parser: change how present items are stored
Instead of remembering the items by index, they are instead remembered
by offset.  This allows other parts of the code to easily query
whether a member was set or not without having access to the list of
resource items.

This works since no two members of a class can have the same offset.