devtools: fix dist-tarball.sh path prefix
dist-tarball.sh was supposed to create "clean" tarballs, that is with
a path-prefix of bareos-x.y.z based on the version-number. This didn't
happen as rewriting in tar did not take place.
With this patch the paths provided to tar are now all prefixed with './'
which allows the substitution to work correctly.